pratt: make all states nameless public
